show movie in static pages
