+++ app/controllers/static_pages_controller.rb
@@ -1,6 +1,5 @@
 class StaticPagesController < ApplicationController
   def home
-    @pagy, @movie = pagy Movie.all.sort_list,
-                               items: Settings.digits.movie_per_page
+    @categories = Category.all
   end
 end
