+++ app/javascript/application.js
@@ -1,5 +1,5 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
-import "@hotwired/turbo-rails"
+// import "@hotwired/turbo-rails"
 import "jquery"
 import "jquery_ujs"
 import "popper"
@@ -14,6 +14,7 @@ import "mixi"
 import "player"
 import "nice-select"
 import "controllers"
+import "window-load"
 import "custom"
 import "admin"
 import "main-layout"
