+++ app/views/movies/_movie_index.html.erb
@@ -1,0 +1,20 @@
+<div class="col-lg-4 col-md-6 col-sm-6">
+  <div class="product__item">
+    <% if movie.image.attached? %>
+      <div class="product__item__pic set-bg"  data-setbg=<%= url_for movie.image %>>
+    <% else %>
+      <div class="product__item__pic set-bg" data-setbg=<%= movie.img_link %>>
+    <% end %>
+      <div class="ep">18 / 18</div>
+      <div class="comment"><i class="fa fa-comments"></i> 11</div>
+      <div class="view"><i class="fa fa-eye"></i> 9141</div>
+    </div>
+    <div class="product__item__text">
+      <ul>
+        <li>Active</li>
+        <li>Movie</li>
+      </ul>
+      <h5><%= link_to movie.title, movie_path(id: movie.id) %></h5>
+    </div>
+  </div>
+</div>
