+++ app/views/movies/show.html.erb
@@ -51,7 +51,7 @@
               <div class="row">
                 <div class="col-lg-6 col-md-6">
                   <ul>
-                    <li><span><%= t ".release" %></span><%= l(@movie.release_time, format: :own) %></li>
+                    <li><span><%= t ".release" %></span><%= l(@movie.release_time, format: :default) %></li>
                     <li><span><%= t ".director" %></span><%= @movie.director %></li>
                     <li><span><%= t ".cast" %></span><%= @movie.cast %></li>
                     <li><span><%= t ".status" %></span><%= @movie.status %></li>
