+++ app/views/static_pages/_category_index.html.erb
@@ -1,0 +1,17 @@
+<div class="trending__product">
+  <div class="row">
+    <div class="col-lg-8 col-md-8 col-sm-8">
+      <div class="section-title">
+        <h4><%= category.name %></h4>
+      </div>
+    </div>
+    <div class="col-lg-4 col-md-4 col-sm-4">
+      <div class="btn__all">
+        <a href="#" class="primary-btn">View All <span class="arrow_right"></span></a>
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <%= render partial: "movies/movie_index", collection: category.movies.limit(6), as: :movie %>
+  </div>
+</div>
