+++ config/locales/en.yml
@@ -5,7 +5,7 @@ en:
   time:
     formats:
       own: "%H:%M:%Ss"
-      default: "%d/%m/%Y"
+      default: "%d-%m-%Y"
   text:
     booking_system: Winter Cinema
   layouts:
