+++ db/migrate/20230323132210_add_references.rb
@@ -8,6 +8,5 @@ class AddReferences < ActiveRecord::Migration[7.0]
     add_reference(:tickets, :payment, foreign_key: true)
     add_reference(:show_times, :room, foreign_key: true)
     add_reference(:rooms, :cinema, foreign_key: true)
-    add_reference(:movies, :category, foreign_key: true)
   end
 end
